+++ kamehouse-desktop/bin/widgets/hostname.py
@@ -1,7 +1,0 @@
-import socket
-
-from widgets.text import TextWidget
-
-class HostnameWidget(TextWidget):
-    def __init__(self, window):
-        super().__init__('hostname_widget', socket.gethostname(), window)