+++ kamehouse-desktop/bin/widgets/kamehouse_katakana.py
@@ -1,5 +1,0 @@
-from widgets.text import TextWidget
-
-class KameHouseKatakanaWidget(TextWidget):
-    def __init__(self, window):
-        super().__init__('kamehouse_katakana_widget', "カメハウス", window)