+++ kamehouse-desktop/bin/widgets/kamehouse_logo.py
@@ -1,5 +1,0 @@
-from widgets.image import ImageWidget
-
-class KameHouseLogoWidget(ImageWidget):
-    def __init__(self, window):
-        super().__init__('kamehouse_logo_widget', window)