+++ kamehouse-desktop/bin/widgets/world_cup_logo.py
@@ -1,5 +1,0 @@
-from widgets.image import ImageWidget
-
-class WorldCupLogoWidget(ImageWidget):
-    def __init__(self, window):
-        super().__init__('world_cup_logo_widget', window)